Fix Carbon string type errors + add debug logging for recommendations
- Cast cooldown_days to int in User::handleVerificationRejection() — env()
  returns string which breaks Carbon::addDays() on strict PHP/Carbon
- Cast boost_duration to int in ProfileBoost::activate()
- Add temporary debug logging to handleSearchWithFilters to diagnose
  why swiper shows no users

+++ app/Models/ProfileBoost.php
@@ -51,7 +51,7 @@ class ProfileBoost extends Model
         }
 
         // Get boost duration from the package, default to 30 minutes if not found
-        $boostDuration = $this->boostPackage ? $this->boostPackage->boost_duration : 30;
+        $boostDuration = (int) ($this->boostPackage ? $this->boostPackage->boost_duration : 30);
 
         $this->status = 'active';
         $this->activated_at = now();
